Deleting duplicated AnimationActivity class
