+++ appGame/src/main/java/pg/autyzm/graprzyjazneemocje/animation/AnimationEndActivity.java
@@ -1,0 +1,69 @@
+package pg.autyzm.graprzyjazneemocje.animation;
+
+import android.app.Activity;
+import android.content.Intent;
+import android.os.Bundle;
+import android.view.animation.Animation;
+import android.view.animation.AnimationUtils;
+import android.widget.ImageView;
+
+import pg.autyzm.graprzyjazneemocje.R;
+
+//import AnimationActivity;
+
+/**
+ * Created by Joanna on 2016-12-01.
+ */
+
+public class AnimationEndActivity extends AnimationActivity{
+
+    protected ImageView animImage;
+    protected Animation anim;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+        randomAward();
+
+        anim.setAnimationListener(this);
+        animImage.startAnimation(anim);
+
+    }
+
+    @Override
+    public void onAnimationEnd(Animation animation) {
+        Intent resultIntent = new Intent();
+        setResult(Activity.RESULT_OK, resultIntent);
+        finish();
+    }
+
+    @Override
+    public void onAnimationStart(Animation animation) {
+    }
+
+    @Override
+    public void onAnimationRepeat(Animation animation) {
+    }
+
+
+    protected void randomAward() {
+        setContentView(R.layout.activity_anim_end);
+
+        int offset = -300;
+        int butterflyimages[] = {R.id.butterfly1_image, R.id.image2, R.id.image3, R.id.image4, R.id.butterfly5_image};
+
+        for (int image : butterflyimages) {
+            animImage = (ImageView) findViewById(image);
+            anim = AnimationUtils.loadAnimation(getApplicationContext(), R.anim.spiralend);
+            anim.setStartOffset(offset += 300);
+            animImage.startAnimation(anim);
+        }
+
+
+        animImage = (ImageView) findViewById(R.id.sun_image);
+        anim = AnimationUtils.loadAnimation(getApplicationContext(), R.anim.sun);
+        animImage.startAnimation(anim);
+
+    }
+}
